cgofuse: linting + reorganizing
