+++ internal/filesystem/cgofuse/dirstat_other.go
@@ -1,0 +1,16 @@
+// go:build !windows
+
+package cgofuse
+
+import (
+	"io/fs"
+
+	"github.com/winfsp/cgofuse/fuse"
+)
+
+// [2022.11.15] readdir-plus in cgofuse is only supported on Windows.
+// If support for a system is added in cgofuse,
+// metadata should be returned within `readdir` in this project as well.
+// This function is a no-op since FUSE will use `getattr` instead
+// to retrieve metadata on systems without the readdir-plus capability.
+func dirStat(fs.DirEntry) *fuse.Stat_t { return nil }
