+++ internal/filesystem/cgofuse/wrapper.go
@@ -2,70 +2,20 @@ package cgofuse
 
 import (
 	"io/fs"
-	"log"
-	"math"
 
-	fuselib "github.com/winfsp/cgofuse/fuse"
-	"github.com/djdv/go-filesystem-utils/internal/filesystem/errors"
 	"github.com/u-root/uio/ulog"
+	"github.com/winfsp/cgofuse/fuse"
 )
 
 const (
-	operationSuccess = 0
-	errorHandle      = math.MaxUint64
-	handleMax        = errorHandle - 1
-
-	S_IRWXO = fuselib.S_IROTH | fuselib.S_IWOTH | fuselib.S_IXOTH
-	S_IRWXG = fuselib.S_IRGRP | fuselib.S_IWGRP | fuselib.S_IXGRP
-	S_IRWXU = fuselib.S_IRUSR | fuselib.S_IWUSR | fuselib.S_IXUSR
-
-	IRWXA = S_IRWXU | S_IRWXG | S_IRWXO                                    // 0o777
-	IRXA  = IRWXA &^ (fuselib.S_IWUSR | fuselib.S_IWGRP | fuselib.S_IWOTH) // 0o555
+	goRoot = "."
 )
 
 type goWrapper struct {
 	fileTable
 	systemLock operationsLock
 	fs.FS
-	log ulog.Logger // general operations log
-}
-
-const (
-	posixRoot = "/"
-	goRoot    = "."
-)
-
-// FUSE absolute path to relative [fs.FS] name.
-func fuseToGo(path string) (name string, _ error) {
-	const op errors.Op = "path lexer"
-	switch path {
-	case "":
-		return "", errors.New(op,
-			errors.Path("{empty-string}"),
-			errors.InvalidItem,
-		)
-	case posixRoot:
-		return goRoot, nil
-	}
-
-	// TODO: does fuse guarantee slash prefixed paths?
-	return path[1:], nil
-}
-
-type fuseFileType = uint32
-
-// [FileMode] to FUSE mode bits.
-func goToFuseFileType(m fs.FileMode) fuseFileType {
-	switch m.Type() {
-	case fs.ModeDir:
-		return fuselib.S_IFDIR
-	case fs.FileMode(0):
-		return fuselib.S_IFREG
-	case fs.ModeSymlink:
-		return fuselib.S_IFLNK
-	default:
-		return 0
-	}
+	log ulog.Logger
 }
 
 func (fs *goWrapper) Init() {
@@ -75,15 +25,9 @@ func (fs *goWrapper) Init() {
 	defer fs.log.Print("Init finished")
 }
 
-func (fuse *goWrapper) Getattr(path string, stat *fuselib.Stat_t, fh uint64) int {
+func (fuse *goWrapper) Getattr(path string, stat *fuse.Stat_t, fh uint64) int {
 	defer fuse.systemLock.Access(path)()
 	fuse.log.Printf("Getattr - {%X}%q", fh, path)
-
-	defer func() {
-		if r := recover(); r != nil {
-			log.Printf("🔥 HOT TONIGHT - %s 🔥 %v", path, r)
-		}
-	}()
 
 	goPath, err := fuseToGo(path)
 	if err != nil {
@@ -101,7 +45,7 @@ func (fuse *goWrapper) Getattr(path string, stat *fuselib.Stat_t, fh uint64) int
 	if err != nil {
 		errNo := interpretError(err)
 		// Don't flood the logs with "not found" errors.
-		if errNo != -fuselib.ENOENT {
+		if errNo != -fuse.ENOENT {
 			fuse.log.Print(err)
 		}
 		return errNo
@@ -111,9 +55,9 @@ func (fuse *goWrapper) Getattr(path string, stat *fuselib.Stat_t, fh uint64) int
 	// push changes back to fs.FS via extension
 	// fs.SetAttr, fs.SetAttrFuse(path, someOverlappingAttrType)
 
-	mTime := fuselib.NewTimespec(goStat.ModTime())
+	mTime := fuse.NewTimespec(goStat.ModTime())
 
-	stat.Uid, stat.Gid, _ = fuselib.Getcontext()
+	stat.Uid, stat.Gid, _ = fuse.Getcontext()
 	stat.Mode = goToFuseFileType(goStat.Mode()) |
 		IRXA // TODO: permissions from root <- options <- cli
 		// TODO: mask <- check spec; does fuse need one or does it apply one itself?
